% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromDrawFunctions.R
\name{draw.pairwise}
\alias{draw.pairwise}
\title{Pairwise synteny plot}
\usage{
draw.pairwise(infile, output, refSizes, tarSizes, refName, tarName)
}
\arguments{
\item{infile}{Path to the syntenic blocks file}

\item{output}{file name}

\item{refSizes}{file with chromosome sizes for the reference species}

\item{tarSizes}{file with the chromosome sizes for the target species}

\item{refName}{name of the reference species}

\item{tarName}{name of the target species}
}
\value{
A pdf file with the comparative drawings
}
\description{
Draw Pairwise Synteny Plots
}
\details{
This function draws pairwise synteny plots. It requieres as input the syntenic blocks following this
format: referenceSpecies,chr,start,end,targetChr,targetStart,targetEnd,orient,targetSpecies. Please separate bu tabs.
It also requieres the output file name and the range of chromosomes of the reference and target species.
Example: draw.pairwise(input.txt,outputName, refSizes, tarSizes, refName, tarName)
}
