% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromDrawFunctions.R
\name{reformat.syntenyData}
\alias{reformat.syntenyData}
\title{Reformat synteny data}
\usage{
reformat.syntenyData(
  file_data,
  filename,
  reference.species = reference.sps,
  target.species = target.sps
)
}
\arguments{
\item{file_data}{input file name for descrambler .map data}

\item{filename}{output file name for reformatted data}

\item{reference.species}{reference species identifier as a character string}

\item{target.species}{target species identifier as a character string}
}
\value{
A text file in the working directory with the reformatted data
}
\description{
Reformat synteny data
}
\details{
This function takes output from alignment softwares such as deschrambler and inferCARs and reformats it for syntenyPlotteR - this does not curate files only reformats it

It requires as input:

1. The data output from deschrambler or inferCARs

2. the output file name

There are optional parameters for some customization of this function:

1. reference.species allows you to set the reference species identifier that will be set in the final output table i.e. reference.species = "ref"

2. target.species allows you to set the target species identifier that will be set in the final output table i.e. target.species = "tar"

The following example can be recreated with the example data files found in (https://github.com/Farre-lab/syntenyPlotteR/blob/master/data)

Example: reformat.syntenyData("deschrambler.output","reformatted.data",reference.species ="ref", target.species = "tar" )
}
