% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromDrawFunctions.R
\name{draw.eh}
\alias{draw.eh}
\title{Evolution Highway style plot}
\usage{
draw.eh(infile, output, chrRange)
}
\arguments{
\item{infile}{Path to the syntenic blocks file}

\item{output}{file name}

\item{chrRange}{range of chromosome numbers in the reference "1:29"}
}
\value{
A pdf file with the comparative drawings
}
\description{
Draw Evolution Highway Plots
}
\details{
This function draws Evolution Highway style plots. It requieres as input the syntenic blocks following this
format: chr,start,end,targetChr,targetStart,targetEnd,orient,targetSpecies separated by tabs
It also requieres the output file name and the range of chromosomes of the reference species.
Example: draw.eh(input.csv,goat,"1:29")
}
