% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chromDrawFunctions.R
\name{draw.ideogram}
\alias{draw.ideogram}
\title{Draw ideograms in inferCARs style}
\usage{
draw.ideogram(file_data, file_tarsize, file_refsize)
}
\arguments{
\item{file_data}{Path to the syntenic blocks file}

\item{file_tarsize}{Path to the target chromosomes length file}

\item{file_refsize}{Path to the reference chromosomes length file}
}
\value{
A pdf file with the ideogram
}
\description{
Draw synteny ideograms in inferCARS style
}
\details{
This function draws pairwise synteny plots in inferCARS style.
Inputs are tab separated files, one with the synteny blocks and two files with target and reference chromosome sizes.
Synteny blocks file should be as: targetChr targetStart targetEnd referenceChr referenceStart referenceEnd Orientation
Output will be a pdf file with the ideogram.

Target is the species which chromosomes will be painted. Reference will be used for painting and diagonals.
Chromosomes will be in the same order as in the target sizes file.

Example: draw_ideogram(synteny_file, target_chr_size, reference_chr_size)
}
