% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SyntenyPlotteR-package.R
\docType{package}
\name{syntenyPlotteR-package}
\alias{syntenyPlotteR}
\alias{syntenyPlotteR-package}
\title{syntenyPlotteR: Genome Synteny Visualization}
\description{
Draw syntenic relationships between genome assemblies. There are 3 functions which take a tab delimited file containing alignment data for syntenic blocks between genomes to produce either a linear alignment plot, an evolution highway style plot, or a painted ideogram representing syntenic relationships. There is also a function to convert alignment data in the DESCHRAMBLER/inferCAR format to the required data structure.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/Farre-lab/syntenyPlotteR/issues}
}

}
\author{
\strong{Maintainer}: Marta Farre \email{m.farre-belmonte@kent.ac.uk} [copyright holder]

Authors:
\itemize{
  \item Sarah Quigley \email{slq4@kent.ac.uk}
  \item Joana Damas \email{joanadamas@gmail.com}
  \item Denis Larkin \email{dlarkin@rvc.ac.uk}
}

}
\keyword{internal}
